+++ packages/cursorless-engine/src/CommandRunner.ts
@@ -1,5 +1,4 @@
-import type { CommandComplete } from "@cursorless/common";
-import type { CommandResponse } from ".";
+import type { CommandComplete, CommandResponse } from "@cursorless/common";
 
 export interface CommandRunner {
   run(command: CommandComplete): Promise<CommandResponse>;
